* Pg sync patch.  Storage / Ygg now works with Pg again.
* Storage/Structure.pm: Allow NULLs for w/r/m in
permissions.  This is a side effect of having ticket #15
still being unfixed.  As we create the initial permissions
for a role and only want to set 'r', we're leaving 'm' and
'w' as undef.  The fix is probably to read up the old
values and set the new value issued to the grant while
retaining the old and unchanged values as they were.

* Storage.pm: we're not allowed to set NULL to the id / counter
field in the idcounter.  This leaves Store key-less.  This code
path should be validated and ensured to work.