* Implemented search() for the remote API.
* Temporality is still very iffy via the remote API.

* There's some serious hacking of structures on the recieving side of
search in the client API, mostly because the data structures we can
return are somewhat limited.  Complex structures don't work well, so
we flatten the return values in the Client API and fix it in Remote.pm.

This isn't "pretty" but it is at least commented in the code.  Also,
outside search() we probably won't have to deal with this case again.