* Instance creation works.  It's done from the Entity page, and upon creating you're forwarded to the page for the new instance. Creation is done in Instance.pm.
* Redid all forms to manual code to avoid CGI.pms habit of retaining
params.  This is a fugly fix, but I couldn't get CGI.pm to play ball
and just ended up writing the code myself.

* Property creation works.  It's done from the Entity page and
upon creating the property, you're directed back to the Entity page
which now has the new Property added.  Creation is as such done
in Entity.pm.  There is no Property module, as it'd be the most boring
web page ever -- unless someone figures out something fun to do with it.

* We could use some elegant popup saying "data has been stored" or somesuch
upon writes, just to ensure that the user is appropriately informed about
what has happened.