* Added a query cache.
* The query cache expects statements and optional parameters as its keys.

* The query cache is a single instance contained within each Storage object.
Running concurrent Storage objects on the same backend is a bad idea. In the
future, a way to disable the cache (at a large performance penalty) might
be offered.

* When setting a statement, the issuer of the cache value also has to provide
the schemas that this statement depends upon.  If a schema in this list is
written to, all statements using the given schema is expunged from the cache.

* There is currently no size limit to the cache. Whops.