* Passed the storage object to the Querycache.
* Made the Querycache adhere to the debug API.

* The debug interface now has a 'verbose' option, this will print out
some information about the run when the ends.  By default the information
returned is the amount of time the debug option was enabled, but the
user can use debugger->activity( 'keyword' ) (see below) to increase
output.

* Via the "activity" inteface to the debug object, the user can count
occurances of a given event (via the keyword supplied).  At the
end of the run (either when the debug option is set to 0 or the 
application ends) and the verbose option is set, Yggdrasil will
display the activity data to STDERR.